requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.17.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.2.4
grpcio==1.73.0
hf-xet==1.2.0
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.9
lightning==2.5.2
lightning-bolts==0.7.0
lightning-utilities==0.14.3
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.5.0
munkres==1.1.4
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
omegaconf==2.3.0
opencv-python==4.12.0.88
optuna==4.6.0
optuna-integration==4.6.0
packaging==25.0
pandas==2.3.0
pillow==11.0.0
platformdirs==4.9.2
propcache==0.3.2
protobuf==6.31.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==1.9.5
pytorch-lightning-bolts==0.3.2.post1
pytz==2025.2
PyYAML==6.0.2
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.0
scipy==1.15.3
sentry-sdk==2.53.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.3.0
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchmetrics==1.7.3
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.57.6
triton==3.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
wandb==0.25.0
Werkzeug==3.1.3
yarl==1.20.1
